renaming viewmodel folder
